+++ Core/Inc/stm32f4xx_it.h
@@ -53,6 +53,7 @@ void BusFault_Handler(void);
 void UsageFault_Handler(void);
 void DebugMon_Handler(void);
 void SysTick_Handler(void);
+void TIM1_TRG_COM_TIM11_IRQHandler(void);
 void USART1_IRQHandler(void);
 void USART2_IRQHandler(void);
 void DMA2_Stream0_IRQHandler(void);